--- a/modules/mockk/src/jvmTest/kotlin/io/mockk/junit5/MockKExtensionTest.kt
+++ b/modules/mockk/src/jvmTest/kotlin/io/mockk/junit5/MockKExtensionTest.kt
@@ -102,5 +102,4 @@ class MockKExtensionTest {
 
         verify { carSpy.relaxedTest() }
     }
-
 }